+++ SFML-game/libs/SFML/Collision.cpp
@@ -5,7 +5,7 @@
  */
 
 #include "stdafx.h"
-#include <SFML/Collision.hpp>
+#include "SFML/Collision.hpp"
 
 namespace sf::Collision {
 
@@ -68,19 +68,19 @@ bool PixelPerfectTest(const sf::Sprite& Object1, const sf::Sprite& Object2, sf::
 		sf::Uint8* mask2 = Bitmasks.GetMask(Object2.getTexture());
 
 		// Loop through our pixels
-		for (int i = Intersection.left; i < Intersection.left+Intersection.width; i++) {
-			for (int j = Intersection.top; j < Intersection.top+Intersection.height; j++) {
+		for (int i = static_cast<int>(Intersection.left); i < static_cast<int>(Intersection.left+Intersection.width); i++) {
+			for (int j = static_cast<int>(Intersection.top); j < static_cast<int>(Intersection.top+Intersection.height); j++) {
  
-				sf::Vector2f o1v = Object1.getInverseTransform().transformPoint(i, j);
-				sf::Vector2f o2v = Object2.getInverseTransform().transformPoint(i, j);
+				sf::Vector2f o1v = Object1.getInverseTransform().transformPoint(static_cast<float>(i), static_cast<float>(j));
+				sf::Vector2f o2v = Object2.getInverseTransform().transformPoint(static_cast<float>(i), static_cast<float>(j));
  
 				// Make sure pixels fall within the sprite's subrect
 				if (o1v.x > 0 && o1v.y > 0 && o2v.x > 0 && o2v.y > 0 &&
 					o1v.x < O1SubRect.width && o1v.y < O1SubRect.height &&
 					o2v.x < O2SubRect.width && o2v.y < O2SubRect.height) {
 
-						if (Bitmasks.GetPixel(mask1, Object1.getTexture(), (int)(o1v.x)+O1SubRect.left, (int)(o1v.y)+O1SubRect.top) > AlphaLimit &&
-							Bitmasks.GetPixel(mask2, Object2.getTexture(), (int)(o2v.x)+O2SubRect.left, (int)(o2v.y)+O2SubRect.top) > AlphaLimit)
+						if (Bitmasks.GetPixel(mask1, Object1.getTexture(), static_cast<int>(o1v.x)+O1SubRect.left, static_cast<int>(o1v.y)+O1SubRect.top) > AlphaLimit &&
+							Bitmasks.GetPixel(mask2, Object2.getTexture(), static_cast<int>(o2v.x)+O2SubRect.left, static_cast<int>(o2v.y)+O2SubRect.top) > AlphaLimit)
 							return true;
 
 				}
